+++ drivers/clock_control/Kconfig.stm32
@@ -65,7 +65,8 @@ config CLOCK_STM32_MCO1_SRC_LSE
 	bool "LSE"
 	depends on SOC_SERIES_STM32F4X || \
 	           SOC_SERIES_STM32F7X || \
-	           SOC_SERIES_STM32L4X
+	           SOC_SERIES_STM32L4X || \
+	           SOC_SERIES_STM32H7X
 	help
 	  Use LSE as source of MCO1
 
@@ -74,7 +75,8 @@ config CLOCK_STM32_MCO1_SRC_HSE
 	depends on SOC_SERIES_STM32F1X || \
 	           SOC_SERIES_STM32F4X || \
 	           SOC_SERIES_STM32F7X || \
-	           SOC_SERIES_STM32L4X
+	           SOC_SERIES_STM32L4X || \
+	           SOC_SERIES_STM32H7X
 	help
 	  Use HSE as source of MCO1
 
@@ -92,7 +94,10 @@ config CLOCK_STM32_MCO1_SRC_MSI
 
 config CLOCK_STM32_MCO1_SRC_HSI
 	bool "HSI"
-	depends on SOC_SERIES_STM32F1X || SOC_SERIES_STM32F4X || SOC_SERIES_STM32F7X
+	depends on SOC_SERIES_STM32F1X || \
+	           SOC_SERIES_STM32F4X || \
+	           SOC_SERIES_STM32F7X || \
+	           SOC_SERIES_STM32H7X
 	help
 	  Use HSI as source of MCO1
 
@@ -104,7 +109,8 @@ config CLOCK_STM32_MCO1_SRC_HSI16
 
 config CLOCK_STM32_MCO1_SRC_HSI48
 	bool "HSI48"
-	depends on SOC_SERIES_STM32L4X
+	depends on SOC_SERIES_STM32L4X || \
+	           SOC_SERIES_STM32H7X
 	help
 	  Use HSI48 as source of MCO1
 
@@ -113,6 +119,12 @@ config CLOCK_STM32_MCO1_SRC_PLLCLK
 	depends on SOC_SERIES_STM32F4X || SOC_SERIES_STM32F7X || SOC_SERIES_STM32L4X
 	help
 	  Use PLLCLK as source of MCO1
+
+config CLOCK_STM32_MCO1_SRC_PLLQCLK
+	bool "PLLQ"
+	depends on SOC_SERIES_STM32H7X
+	help
+	  Use PLLQ as source of MCO1
 
 config CLOCK_STM32_MCO1_SRC_PLLCLK_DIV2
 	bool "PLLCLK_DIV2"
@@ -150,10 +162,12 @@ config CLOCK_STM32_MCO1_DIV
 	depends on !CLOCK_STM32_MCO1_SRC_NOCLOCK && (\
 	             SOC_SERIES_STM32F4X || \
 	             SOC_SERIES_STM32F7X || \
-	             SOC_SERIES_STM32L4X \
+	             SOC_SERIES_STM32L4X || \
+	             SOC_SERIES_STM32H7X \
 	           )
 	default 1
 	range 1   5 if SOC_SERIES_STM32F4X || SOC_SERIES_STM32F7X
+	range 1   15 if SOC_SERIES_STM32H7X
 	range 1   16 if SOC_SERIES_STM32L4X
 	help
 	  Prescaler for MCO1 output clock
@@ -169,7 +183,9 @@ config CLOCK_STM32_MCO2_SRC_NOCLOCK
 
 config CLOCK_STM32_MCO2_SRC_SYSCLK
 	bool "SYSCLK"
-	depends on SOC_SERIES_STM32F4X || SOC_SERIES_STM32F7X
+	depends on SOC_SERIES_STM32F4X || \
+	           SOC_SERIES_STM32F7X || \
+	           SOC_SERIES_STM32H7X
 	help
 	  Use SYSCLK as source of MCO2
 
@@ -181,9 +197,23 @@ config CLOCK_STM32_MCO2_SRC_PLLI2S
 
 config CLOCK_STM32_MCO2_SRC_HSE
 	bool "HSE"
-	depends on SOC_SERIES_STM32F4X || SOC_SERIES_STM32F7X
+	depends on SOC_SERIES_STM32F4X || \
+	           SOC_SERIES_STM32F7X || \
+	           SOC_SERIES_STM32H7X
 	help
 	  Use HSE as source of MCO2
+
+config CLOCK_STM32_MCO2_SRC_LSI
+	bool "LSI"
+	depends on SOC_SERIES_STM32H7X
+	help
+	  Use LSI as source of MCO2
+
+config CLOCK_STM32_MCO2_SRC_CSI
+	bool "CSI"
+	depends on SOC_SERIES_STM32H7X
+	help
+	  Use CSI as source of MCO2
 
 config CLOCK_STM32_MCO2_SRC_PLLCLK
 	bool "PLLCLK"
@@ -191,14 +221,30 @@ config CLOCK_STM32_MCO2_SRC_PLLCLK
 	help
 	  Use PLLCLK as source of MCO2
 
+config CLOCK_STM32_MCO2_SRC_PLLPCLK
+	bool "PLLPCLK"
+	depends on SOC_SERIES_STM32H7X
+	help
+	  Use PLLPCLK as source of MC02
+
+config CLOCK_STM32_MCO2_SRC_PLL2PCLK
+	bool "PLL2PCLK"
+	depends on SOC_SERIES_STM32H7X
+	help
+	  Use PLL2PCLK as source of MC02
 endchoice
 
 config CLOCK_STM32_MCO2_DIV
 	int "MCO2 prescaler"
-	depends on !CLOCK_STM32_MCO2_SRC_NOCLOCK
+	depends on !CLOCK_STM32_MCO2_SRC_NOCLOCK && (\
+	             SOC_SERIES_STM32F4X || \
+	             SOC_SERIES_STM32F7X || \
+	             SOC_SERIES_STM32H7X \
+	           )
 	default 1
-	range 1 5
+	range 1 5 if SOC_SERIES_STM32F4X || SOC_SERIES_STM32F7X
+	range 1 15 if SOC_SERIES_STM32H7X
 	help
-	  allowed values: 1, 2, 3, 4, 5
+	  Prescaler for MCO2 output clock
 
 endif # CLOCK_CONTROL_STM32_CUBE
