+++ drivers/clock_control/clock_stm32_ll_common.h
@@ -14,48 +14,6 @@
 
 #include <stm32_ll_utils.h>
 
-#if CONFIG_CLOCK_STM32_MCO1_SRC_NOCLOCK
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_NOCLOCK
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_EXT_HSE
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_EXT_HSE
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_LSE
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_LSE
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_HSE
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_HSE
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_LSI
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_LSI
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_MSI
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_MSI
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_HSI
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_HSI
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_HSI16
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_HSI
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_HSI48
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_HSI48
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_PLLCLK
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_PLLCLK
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_PLLCLK_DIV2
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_PLLCLK_DIV_2
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_PLL2CLK
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_PLL2CLK
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_PLLI2SCLK
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_PLLI2SCLK
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_PLLI2SCLK_DIV2
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_PLLI2SCLK_DIV2
-#elif CONFIG_CLOCK_STM32_MCO1_SRC_SYSCLK
-	#define MCO1_SOURCE		LL_RCC_MCO1SOURCE_SYSCLK
-#endif
-
-#if CONFIG_CLOCK_STM32_MCO2_SRC_SYSCLK
-	#define MCO2_SOURCE		LL_RCC_MCO2SOURCE_SYSCLK
-#elif CONFIG_CLOCK_STM32_MCO2_SRC_PLLI2S
-	#define MCO2_SOURCE		LL_RCC_MCO2SOURCE_PLLI2S
-#elif CONFIG_CLOCK_STM32_MCO2_SRC_HSE
-	#define MCO2_SOURCE		LL_RCC_MCO2SOURCE_HSE
-#elif CONFIG_CLOCK_STM32_MCO2_SRC_PLLCLK
-	#define MCO2_SOURCE		LL_RCC_MCO2SOURCE_PLLCLK
-#endif
-
 /* Macros to fill up multiplication and division factors values */
 #define z_pllm(v) LL_RCC_PLLM_DIV_ ## v
 #define pllm(v) z_pllm(v)
