+++ drivers/clock_control/clock_stm32_ll_h7.c
@@ -16,6 +16,7 @@
 #include <zephyr/drivers/clock_control.h>
 #include <zephyr/sys/util.h>
 #include <zephyr/drivers/clock_control/stm32_clock_control.h>
+#include "clock_stm32_ll_mco.h"
 #include "stm32_hsem.h"
 
 
@@ -835,6 +836,9 @@ int stm32_clock_control_init(const struct device *dev)
 #endif
 
 	z_stm32_hsem_lock(CFG_HW_RCC_SEMID, HSEM_LOCK_DEFAULT_RETRY);
+
+	/* Configure MCO1/MCO2 based on Kconfig */
+	stm32_clock_control_mco_init();
 
 	/* Set up indiviual enabled clocks */
 	set_up_fixed_clock_sources();
